backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
deep-translator==1.11.4
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.112.1
flask==3.0.3
flask-sqlalchemy==3.1.1
googletrans==3.1.0a0
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
jinja2==3.1.4
jose==1.0.0
markupsafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.32
starlette==0.38.2
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
werkzeug==3.0.4